update one ui component for pdf header
