+++ apps/employee-monitoring/src/components/pdf/DetailedReportOnOfficialBusinessPassSlip.tsx
@@ -5,8 +5,8 @@ import React, { FunctionComponent, useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { isEmpty } from 'lodash';
 import { DetailedReportOnObPassSlip } from '../../utils/types/report.type';
-import Header from './Header';
 import { DateFormatter } from 'libs/utils/src/lib/functions/DateFormatter';
+import { PdfHeader } from '@gscwd-apps/oneui';
 
 type DetailedReportOnOfficialBusinessPassSlipProps = {
   detailedReportOnObPassSlipDoc: DetailedReportOnObPassSlip;
@@ -121,7 +121,7 @@ export const DetailedReportOnOfficialBusinessPassSlipPdf: FunctionComponent<
             <Page size={[612.0, 936.0]} style={styles.page}>
               <View>
                 {/* HEADER */}
-                <Header withIsoCode={false} withIsoLogo={false} isFixed={true} />
+                <PdfHeader isFixed={true} />
 
                 {/* DOCUMENT TITLE */}
                 <View style={[styles.w100, styles.horizontalCenter]}>
