+++ apps/employee-monitoring/src/components/pdf/DetailedReportOnPersonalBusinessPassSlip.tsx
@@ -5,8 +5,8 @@ import React, { FunctionComponent, useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { isEmpty } from 'lodash';
 import { DetailedReportOnPbPassSlip } from '../../utils/types/report.type';
-import Header from './Header';
 import { DateFormatter } from 'libs/utils/src/lib/functions/DateFormatter';
+import { PdfHeader } from '@gscwd-apps/oneui';
 
 type DetailedReportOnPersonalBusinessPassSlipProps = {
   detailedReportOnPbPassSlipDoc: DetailedReportOnPbPassSlip;
@@ -121,7 +121,7 @@ export const DetailedReportOnPersonalBusinessPassSlipPdf: FunctionComponent<
             <Page size={[612.0, 936.0]} style={styles.page}>
               <View>
                 {/* HEADER */}
-                <Header withIsoCode={false} withIsoLogo={false} isFixed={true} />
+                <PdfHeader isFixed={true} />
 
                 {/* DOCUMENT TITLE */}
                 <View style={[styles.w100, styles.horizontalCenter]}>
