+++ apps/employee-monitoring/src/components/pdf/ReportOnAttendance.tsx
@@ -5,8 +5,8 @@ import React, { FunctionComponent, useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { isEmpty } from 'lodash';
 import { ReportOnAttendance } from '../../utils/types/report.type';
-import Header from './Header';
 import { DateFormatter } from 'libs/utils/src/lib/functions/DateFormatter';
+import { PdfHeader } from '@gscwd-apps/oneui';
 
 type ReportOnAttendancePdfProps = {
   reportOnAttendanceData: ReportOnAttendance;
@@ -116,7 +116,7 @@ export const ReportOnAttendancePdf: FunctionComponent<ReportOnAttendancePdfProps
             <Page size={[612.0, 936.0]} style={styles.page}>
               <View>
                 {/* HEADER */}
-                <Header withIsoCode={false} withIsoLogo={false} isFixed={true} />
+                <PdfHeader isFixed={true} />
 
                 {/* DOCUMENT TITLE */}
                 <View style={[styles.w100, styles.horizontalCenter]}>
